+++ src/TraceEvent/TraceEvent.Tests/inputs/eventpipe-dotnetcore2.0-win-x86.netperf.baseline.txt
@@ -1,0 +1,18 @@
+Microsoft-DotNETCore-SampleProfiler/Thread/Sample, 1, <Event MSec=     "2.4754" PID=   "0" PName=        "" TID="3872" EventName="Thread/Sample" Type="External">\r\n</Event>
+Microsoft-Windows-DotNETRuntime/AppDomainResourceManagement/ThreadCreated, 1, <Event MSec=     "2.3211" PID=   "0" PName=        "" TID="15224" EventName="AppDomainResourceManagement/ThreadCreated" ManagedThreadID="0x00B511C8" AppDomainID="0x00AF9A90" Flags="0x00000000" ManagedThreadIndex="3" OSThreadID="15,224" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/AllocationTick, 1, <Event MSec=     "5.9693" PID=   "0" PName=        "" TID="3872" EventName="GC/AllocationTick" AllocationAmount="0x000192B4" AllocationKind="Small" ClrInstanceID="2" AllocationAmount64="103,092" TypeID="0x09EE4E84" TypeName="System.Object" HeapIndex="0" Address="0x06E23FE4"/>
+Microsoft-Windows-DotNETRuntime/GC/GlobalHeapHistory, 10, <Event MSec=     "8.7952" PID=   "0" PName=        "" TID="3872" EventName="GC/GlobalHeapHistory" FinalYoungestDesired="0x00200000" NumHeaps="1" CondemnedGeneration="2" Gen0ReductionCount="0" Reason="Induced" GlobalMechanisms="Compaction, Promotion" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/HeapStats, 10, <Event MSec=     "8.8868" PID=   "0" PName=        "" TID="3872" EventName="GC/HeapStats" TotalPromoted="0x0001C1C8" TotalHeapSize="0x0001C2C8" Depth="0x00000000" GenerationSize0="0x0000000C" TotalPromotedSize0="0x00017FB8" GenerationSize1="0x00018060" TotalPromotedSize1="0x00000000" GenerationSize2="0x0000000C" TotalPromotedSize2="0x00000000" GenerationSize3="0x00004250" TotalPromotedSize3="0x00004210" FinalizationPromotedSize="0x00000000" FinalizationPromotedCount="0" PinnedObjectCount="5" SinkBlockCount="1" GCHandleCount="27" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/MarkWithType, 30, <Event MSec=     "6.6881" PID=   "0" PName=        "" TID="3872" EventName="GC/MarkWithType" HeapNum="0" ClrInstanceID="2" Type="0" Promoted="390"/>
+Microsoft-Windows-DotNETRuntime/GC/PerHeapHistory, 10, <Event MSec=     "8.8299" PID=   "0" PName=        "" TID="3872" EventName="GC/PerHeapHistory" ClrInstanceID="2" FreeListAllocated="0" FreeListRejected="0" EndOfSegAllocated="0" CondemnedAllocated="0" PinnedAllocated="0" PinnedAllocatedAdvance="0" RunningFreeListEfficiency="0" CondemnReasons0="10" CondemnReasons1="8,193" CompactMechanisms="compact_low_ephemeral" ExpandMechanisms="not_specified" HeapIndex="0" ExtraGen0Commit="380,808" Count="4"/> <GenData Name="Gen0" SizeBefore="446,600" SizeAfter="12" ObjSpaceBefore="446,600" Fragmentation="0" ObjSizeAfter="12" FreeListSpaceBefore="0" FreeObjSpaceBefore="0" FreeListSpaceAfter="0" FreeObjSpaceAfter="0" In="0" Out="98,232" NewAllocation="2,097,152" SurvRate="21" PinnedSurv="28" NonePinnedSurv="98,204"/>\r\n <GenData Name="Gen1" SizeBefore="12" SizeAfter="98,400" ObjSpaceBefore="12" Fragmentation="12" ObjSizeAfter="98,388" FreeListSpaceBefore="0" FreeObjSpaceBefore="0" FreeListSpaceAfter="0" FreeObjSpaceAfter="12" In="98,232" Out="0" NewA
+Microsoft-Windows-DotNETRuntime/GC/PinObjectAtGCTime, 50, <Event MSec=     "6.9443" PID=   "0" PName=        "" TID="3872" EventName="GC/PinObjectAtGCTime" HandleID="0x005D12EC" ObjectID="0x07DC3260" ObjectSize="8,172" TypeName="System.Object[]" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/RestartEEStart, 11, <Event MSec=     "2.4880" PID=   "0" PName=        "" TID="15224" EventName="GC/RestartEEStart" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/RestartEEStop, 11, <Event MSec=     "2.5204" PID=   "0" PName=        "" TID="15224" EventName="GC/RestartEEStop" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/SuspendEEStart, 11, <Event MSec=     "2.3428" PID=   "0" PName=        "" TID="15224" EventName="GC/SuspendEEStart" Reason="SuspendOther" Count="-1" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/SuspendEEStop, 11, <Event MSec=     "2.3799" PID=   "0" PName=        "" TID="15224" EventName="GC/SuspendEEStop" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/GC/Triggered, 10, <Event MSec=     "6.5033" PID=   "0" PName=        "" TID="3872" EventName="GC/Triggered" Reason="Induced" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/Method/JittingStarted, 2, <Event MSec=     "3.8104" PID=   "0" PName=        "" TID="3872" EventName="Method/JittingStarted" MethodID="0x00A95038" ModuleID="0x00A9447C" MethodToken="0x06000004" MethodILSize="0x00000012" MethodNamespace="Tracing.Tests.EventPipeTrace" MethodName=".cctor" MethodSignature="void  ()" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntime/Method/LoadVerbose, 2, <Event MSec=     "5.2382" PID=   "0" PName=        "" TID="3872" EventName="Method/LoadVerbose" MethodID="0x00A95038" ModuleID="0x00A9447C" MethodStartAddress="0x01355138" MethodSize="0x00000028" MethodToken="0x06000004" MethodFlags="Jitted" MethodNamespace="Tracing.Tests.EventPipeTrace" MethodName=".cctor" MethodSignature="void  ()" ClrInstanceID="2" ReJITID="0x00000000"/>
+Microsoft-Windows-DotNETRuntimeRundown/Method/DCStop, 14259, <Event MSec=    "45.1216" PID=   "0" PName=        "" TID="3872" EventName="Method/DCStop" MethodID="0x090B9E28" ModuleID="0x08F41000" MethodStartAddress="0x099054B8" MethodSize="0x000000AF" MethodToken="0x06000001" MethodFlags="None" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntimeRundown/Method/DCStopComplete, 1, <Event MSec= "14639.7539" PID=   "0" PName=        "" TID="3872" EventName="Method/DCStopComplete" ClrInstanceID="2"/>
+Microsoft-Windows-DotNETRuntimeRundown/Method/DCStopInit, 1, <Event MSec=    "44.7592" PID=   "0" PName=        "" TID="3872" EventName="Method/DCStopInit" ClrInstanceID="2"/>
